TD-937 Breaks up the PlatformReportsController class into partial classes for better separation
